📦 chore(pubspec.yaml): update dependencies 📄 feat(global.dart): add global.dart file 📄 feat(models/data.dart): add data.dart file 📄 feat(view/home_page.dart): add home_page.dart file 📄 feat(view/login_page.dart): add login_page.dart file 📄 feat(view/register_page.dart): add register_page.dart file
The dependencies in the pubspec.yaml file have been updated to the latest versions. The new dependencies added are:
- dio: ^5.3.4
- flutter_svg: ^2.0.9
- flutter_bloc: ^8.1.3
- get: ^4.6.6
- shared_preferences: ^2.2.2

The global.dart, data.dart, home_page.dart, login_page.dart, and register_page.dart files have been added to the project. These files are necessary for the implementation of new features or functionality.